GDB-7684: Support expression values for mongo query configs
Added support in the MongoResultIterator to detect when any of the query components
are constructed with a function and not present during the first pass of the statements.
In this case the initial iterator would not be allowed to be closed and will be reused
when the variables are resolved and passed to the configuration.
Any of the following parameters could be set via an expression: query, hint, projection, collation, aggregation.