Add simple batch document processing.
Added functionality where all documents matching the input query, up to a maximum are loaded in memory and returned as named graph collection instead of processing one by one.
When the new functionality is used the plugin is no longer streaming but will require a buffer to store the documents until the query is complete.

Added system property configuration that controls the maximum allowed batch size: graphdb.mongodb.maxBatchSize. This is to prevent OOM problems.