GDB-11761: Add mongo connector batch processing (#24)
* Change plugin Java compatibility to Java 21 to match GDB 11

* Add simple batch document processing.

Added functionality where all documents matching the input query, up to a maximum are loaded in memory and returned as named graph collection instead of processing one by one.
When the new functionality is used the plugin is no longer streaming but will require a buffer to store the documents until the query is complete.

Added system property configuration that controls the maximum allowed batch size: graphdb.mongodb.maxBatchSize. This is to prevent OOM problems.

* GDB-11761 Adds some test cases and fixes the failing ones

- Fixed the failing tests and added couple of new ones, which should
cover the specific client case.

- Added CHANGELOG file.

- Added some releases and branch information to the README.

---------

Co-authored-by: Antoniy Kunchev <antoniy.kunchev@ontotext.com>